moved most of the page findby content to appropriate classes
