+++ src/test/java/org/tacc/green1/model/account/AccountSidebar.java
@@ -2,17 +2,16 @@ package org.tacc.green1.model.account;
 
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
-import org.tacc.green1.model.base.Modal;
+import org.tacc.green1.model.base.Component;
 import org.tacc.green1.util.XPath;
 
 
-public class AccountSidebar extends Modal implements XPath.AccountSidebar {
+public class AccountSidebar extends Component implements XPath.AccountSidebar {
     @FindBy(xpath = LINK_ADDRESS_BOOK)
     private WebElement addressBookLink;
 
     @FindBy(xpath = LINK_ACCOUNT_INFORMATION)
     private WebElement accountInformationLink;
-
 
     public AddressBookPage gotoAddressBookPage() {
         addressBookLink.click();
