+++ src/test/java/org/tacc/green1/model/base/Component.java
@@ -1,3 +1,5 @@
 package org.tacc.green1.model.base;
 
-public abstract class Component extends Modal {}
+public abstract class Component extends Modal {
+
+}
