+++ src/test/java/org/tacc/green1/model/base/Page.java
@@ -2,86 +2,18 @@ package org.tacc.green1.model.base;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.openqa.selenium.TimeoutException;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
-import org.tacc.green1.model.AdvancedSearchPage;
-import org.tacc.green1.model.LoginPage;
-import org.tacc.green1.model.MainPage;
-import org.tacc.green1.model.RegistrationPage;
-import org.tacc.green1.model.account.AccountPopup;
-import org.tacc.green1.model.cart.Cart;
-import org.tacc.green1.model.menu.HeaderMenu;
+import org.tacc.green1.model.base.components.FooterComponent;
+import org.tacc.green1.model.base.components.HeaderComponent;
 import org.tacc.green1.util.XPath;
 
 
 public abstract class Page<T> extends Modal implements XPath.Page {
     private static final Logger LOG = LogManager.getLogger(Page.class);
 
-    @FindBy(xpath = LINK_LOGO)
-    private WebElement logoLink;
-
-    @FindBy(className = "authorization-link")
-    private WebElement loginLink;
-
-    @FindBy(xpath = LINK_REGISTRATION)
-    private WebElement createAccountLink;
-
-    @FindBy(xpath = LINK_WELCOME_ACCOUNT)
-    private WebElement welcomeAccountLink;
-
-    @FindBy(css = "button.action.switch")
-    private WebElement accountPopupToggle;
-
-    @FindBy(xpath = BUTTON_OPEN_CART)
-    private WebElement openCart;
-
-    @FindBy(xpath = LINK_ADVANCED_SEARCH)
-    private WebElement advancedSearchLink;
-
     @FindBy(xpath = BUTTON_CONFIRM_DELETE)
     private WebElement confirmDeleteButton;
-
-
-    public MainPage gotoMainPage() {
-        logoLink.click();
-        return new MainPage();
-    }
-
-    public LoginPage gotoLoginPage() {
-        loginLink.click();
-        return new LoginPage();
-    }
-
-    public RegistrationPage gotoRegistrationPage() {
-        createAccountLink.click();
-        return new RegistrationPage();
-    }
-
-    public AccountPopup openAccountPopup() {
-        if (!isClientLoggedIn()) {
-            String message = "No user is logged in, check out your method call order";
-            LOG.error(message);
-            throw new IllegalStateException(message);
-        }
-
-        accountPopupToggle.click();
-        return new AccountPopup();
-    }
-
-    public HeaderMenu gotoHeaderMenu() {
-        return new HeaderMenu();
-    }
-
-    public Cart openCart() {
-        openCart.click();
-        return new Cart();
-    }
-
-    public AdvancedSearchPage gotoAdvancedSearchPage() {
-        advancedSearchLink.click();
-        return new AdvancedSearchPage();
-    }
 
     @SuppressWarnings("unchecked")
     public T confirmDelete() {
@@ -93,19 +25,13 @@ public abstract class Page<T> extends Modal implements XPath.Page {
         return (T) this;
     }
 
+    public FooterComponent gotoFooterComponent() {
+        LOG.debug(String.format("Moving into %s", FooterComponent.class));
+        return new FooterComponent();
+    }
 
-    public boolean isClientLoggedIn() {
-        timeout(1);
-        try {
-            timeoutByVisibility(welcomeAccountLink);
-        } catch (TimeoutException e) {
-            //Element not found, not logged in
-            LOG.warn("The user is not logged in", e);
-            return false;
-        }
-
-        var message = welcomeAccountLink.getText();
-        //Manual welcome message check
-        return message.contains("Welcome, ");
+    public HeaderComponent gotoHeaderComponent() {
+        LOG.debug(String.format("Moving into %s", HeaderComponent.class));
+        return new HeaderComponent();
     }
 }
